First part of better upload process
